sandbox - drop support for running file service in renderer
