+++ src/vs/platform/windows/electron-main/windowsMainService.ts
@@ -1294,7 +1294,6 @@ export class WindowsMainService extends Disposable implements IWindowsMainServic
 			zoomLevel: typeof windowConfig?.zoomLevel === 'number' ? windowConfig.zoomLevel : undefined,
 
 			legacyWatcher: this.configurationService.getValue('files.legacyWatcher'),
-			experimentalSandboxedFileService: this.configurationService.getValue('files.experimentalSandboxedFileService'),
 			autoDetectHighContrast: windowConfig?.autoDetectHighContrast ?? true,
 			accessibilitySupport: app.accessibilitySupportEnabled,
 			colorScheme: {
