+++ src/vs/workbench/workbench.desktop.sandbox.main.ts
@@ -33,6 +33,14 @@ import 'vs/workbench/electron-sandbox/desktop.main';
 
 //#region --- workbench services
 
+import { IExtensionService, NullExtensionService } from 'vs/workbench/services/extensions/common/extensions';
+import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
+
+// TODO@bpasero sandbox: remove me when extension host is present
+class SimpleExtensionService extends NullExtensionService { }
+
+registerSingleton(IExtensionService, SimpleExtensionService);
+
 
 //#endregion
